address comments & add encryption UT

+++ dbms/src/Storages/DeltaMerge/File/DMFile.cpp
@@ -113,7 +113,7 @@ DMFilePtr DMFile::restore(
     UInt64 page_id,
     const String & parent_path,
     const DMFileMeta::ReadMode & read_meta_mode,
-    UInt32 meta_version,
+    UInt64 meta_version,
     KeyspaceID keyspace_id)
 {
     auto is_s3_file = S3::S3FilenameView::fromKeyWithPrefix(parent_path).isDataFile();
